# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CPIP Backend API
  version: 1.0.0
  description: Corporate Project Intelligence Platform API
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      summary: List projects
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          $ref: '#/components/responses/Success'
    post:
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpsertRequest'
      responses:
        '201':
          $ref: '#/components/responses/Success'
  /projects/{id}:
    get:
      summary: Get project
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '200':
          $ref: '#/components/responses/Success'
    put:
      summary: Update project
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpsertRequest'
      responses:
        '200':
          $ref: '#/components/responses/Success'
    delete:
      summary: Archive project
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '200':
          $ref: '#/components/responses/Success'
...
  /resources:
    get:
      summary: List resources
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          $ref: '#/components/responses/Success'
    post:
      summary: Create resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreateRequest'
      responses:
        '201':
          $ref: '#/components/responses/Success'
  /resources/allocations:
    get:
      summary: Get resource allocations
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          $ref: '#/components/responses/Success'
...
  /risks:
    get:
      summary: List risks
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          $ref: '#/components/responses/Success'
    post:
      summary: Create risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskUpsertRequest'
      responses:
        '201':
          $ref: '#/components/responses/Success'
  /risks/{id}:
    put:
      summary: Update risk
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskUpsertRequest'
      responses:
        '200':
          $ref: '#/components/responses/Success'
...
  /reports/generate:
    post:
      summary: Generate report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportGenerateRequest'
      responses:
        '202':
          $ref: '#/components/responses/Success'
  /reports:
    get:
      summary: List reports
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          $ref: '#/components/responses/Success'
  /reports/{id}:
    get:
      summary: Get report
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '200':
          $ref: '#/components/responses/Success'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    PathID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Success:
      description: Success response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
...
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          additionalProperties: true
        meta:
          type: object
          additionalProperties: true
      required: [success, data]
...
    ProjectUpsertRequest:
      type: object
      properties:
        portfolio_id: { type: string, format: uuid }
        code: { type: string, minLength: 2, maxLength: 50 }
        name: { type: string, minLength: 2, maxLength: 250 }
        manager_user_id: { type: string, format: uuid }
      required: [portfolio_id, code, name, manager_user_id]
...
    ResourceCreateRequest:
      type: object
      properties:
        employee_code: { type: string, minLength: 2, maxLength: 50 }
        name: { type: string, minLength: 2, maxLength: 150 }
        role_title: { type: string, minLength: 2, maxLength: 100 }
        manager_user_id: { type: string, format: uuid }
      required: [employee_code, name, role_title]
...
    RiskUpsertRequest:
      type: object
      properties:
        project_id: { type: string, format: uuid }
        code: { type: string, minLength: 2, maxLength: 40 }
        title: { type: string, minLength: 2, maxLength: 250 }
        likelihood: { type: integer, minimum: 1, maximum: 5 }
        impact: { type: integer, minimum: 1, maximum: 5 }
        owner_user_id: { type: string, format: uuid }
      required: [project_id, code, title, likelihood, impact, owner_user_id]
...
    ReportGenerateRequest:
      type: object
      properties:
        report_type: { type: string, enum: [executive, portfolio, financial, resource, risk, vendor] }
        format: { type: string, enum: [pdf, xlsx, csv] }
        filters:
          type: object
          additionalProperties: true
      required: [report_type, format]